Avoid possible infinite loop in Metabench.measure
The infinite loop would happen if the command we launched always returned
an error. Because we did not check whether the command had failed, we
would assume that it succeeded and would run it again. Now, if the
command fails, we just stop.

We never encountered this problem before, because the CMake command
itself would always succeed, even when the compilation would fail.
With NMake, we're getting an error when launching CMake, so the
problem popped up.