Change the interpolation mode from 'basis' to 'cardinal'
Doing so will provide a accurate match to the data, while preserving
the desired smoothing behaviour.

A better option might be to provide options for all three
interpolation modes (none, cardinal, B-spline), as the B-spline
interpolation is still desirable under certaincircumstances such as
data with a high amount of noise.

For a reference of all the interpolation modes see:
https://github.com/d3/d3-shape/blob/master/README.md#curves